+++ Flounder-Core/src/waters/water.cpp
@@ -6,7 +6,7 @@ namespace flounder
 	const float water::WAVE_LENGTH = 30.0f;
 	const float water::AMPLITUDE = 0.7f;
 
-	const double water::SQUARE_SIZE = 2.598f;
+	const float water::SQUARE_SIZE = 2.598f;
 	const int water::VERTEX_COUNT = 176;
 
 	const colour water::WATER_COLOUR = colour(0.21f, 0.41f, 0.59f);
